+++ MillimanAccessPortal/MapCommonLib/GlobalFunctions.cs
@@ -179,7 +179,7 @@ namespace MapCommonLib
     /// </summary>
     public enum IssueLogEnum
     {
-        QueuePostProcessing
+        LongRunningSelectionGroupProcessing
     }
 
 }